Backport some stuff in this commit - https://github.com/libretro/fceu-next/commit/aea20914a6734f82a0962c8b3204f54c1d06f671

+++ src/drivers/libretro/fceu/video.c
@@ -55,8 +55,8 @@ int FCEU_InitVirtualVideo(void)
 
    if (sizeof(uint8*) == 4)
    {
-      uint32 m;
-      m = *XBuf;
+      uintptr_t m;
+      m = (uintptr_t)*XBuf;
       m = (4 - m) & 3;
       XBuf += m;
    }
